+++ lib/src/data/models/bucket.dart
@@ -2,29 +2,32 @@ import 'package:motor_flutter/motor_flutter.dart';
 import 'package:motor_flutter/src/platform/motor_flutter_platform_interface.dart';
 
 class Bucket {
-  /// [did] is the Unique identifier of the bucket
-  final String did;
-
   /// [whereIs] is the underlying specification of the bucket
   final WhereIs whereIs;
 
-  bool _isInitialized = false;
+  final _cachedDocs = <SchemaDocument>[];
+
+  DateTime _lastUpdated = DateTime.now();
+
+  bool get _needsRefresh => _lastUpdated.isBefore(DateTime.now().subtract(Duration(seconds: 10)));
+
+  String get did => whereIs.did;
 
   // Constructor
-  Bucket(this.did, this.whereIs) {
+  Bucket(this.whereIs) {
     if (MotorFlutter.isReady) {
-      _resolve();
+      refresh();
     }
   }
 
   /// Creates a new bucket from a [CreateBucketResponse]
   factory Bucket.fromResponse(CreateBucketResponse resp) {
-    return Bucket(resp.did, resp.whereIs);
+    return Bucket(resp.whereIs);
   }
 
   /// Creates a new bucket from a [WhereIs]
   factory Bucket.fromWhereIs(WhereIs whereIs) {
-    return Bucket(whereIs.did, whereIs);
+    return Bucket(whereIs);
   }
 
   /// Creates List of [Bucket] from a [List] of [WhereIs]
@@ -34,18 +37,15 @@ class Bucket {
 
   /// Adds a [SchemaDocument] into the Bucket and returns [bool] indicating success
   Future<bool> addDocument(String label, SchemaDocument doc) async {
-    if (!_isInitialized) {
-      await _resolve();
+    await refresh();
+
+    final newDoc = await doc.upload();
+    if (newDoc == null) {
+      return false;
     }
-    if (!doc.hasCid()) {
-      final newDoc = await doc.upload(label);
-      if (newDoc == null) {
-        return false;
-      }
-      doc = newDoc;
-    }
+
     final item = BucketItem(
-      uri: doc.cid,
+      uri: newDoc.cid,
       name: label,
       type: ResourceIdentifier.CID,
       schemaDid: doc.schemaDid,
@@ -53,27 +53,61 @@ class Bucket {
     return await MotorFlutterPlatform.instance.addBucketObject(did, item);
   }
 
-  /// Retrieves a [SchemaDocument] from the Bucket
-  Future<SchemaDocument?> getDocument(String cid) async {
-    if (!_isInitialized) {
-      await _resolve();
-    }
-    final content = await MotorFlutterPlatform.instance.getBucketObject(did, cid);
-    if (content == null) {
-      return null;
-    }
-    try {
-      final item = BucketItem.fromBuffer(content.item);
-      return item.getSchemaDocument();
-    } catch (e) {
-      return null;
-    }
+  /// Retrieves a [SchemaDocument] from the Bucket by [label]
+  Future<SchemaDocument?> getItem(String label) async {
+    await refresh();
+    final doc = _cachedDocs.firstWhere((element) => element.label == label, orElse: () => SchemaDocument());
+    return doc.hasCid() ? doc : null;
   }
 
   /// Lists all [SchemaDocument]s in the Bucket
-  Future<List<SchemaDocument>> listDocuments() async {
-    if (!_isInitialized) {
-      await _resolve();
+  Future<List<SchemaDocument>> listItems() async {
+    await refresh();
+    return _cachedDocs;
+  }
+
+  /// Deletes a [SchemaDocument] from the Bucket given its [label]
+  Future<bool> removeItem(String label) async {
+    await refresh();
+    final doc = _cachedDocs.firstWhere((element) => element.label == label, orElse: () => SchemaDocument());
+    if (!doc.hasCid()) {
+      return false;
+    }
+    return await MotorFlutterPlatform.instance.removeBucketObject(did, doc.cid);
+  }
+
+  /// ### Fetching a Document
+  ///
+  /// Fetches a document from IPFS using the provided [cid]. Returns the [SchemaDocument] if successful, and null if the document was not found.
+  ///
+  /// ```dart
+  /// final bucket = await MotorFlutter.to.getBucket('did:snr:mXyZ123');
+  /// if (res == null) {
+  ///   throw Exception('Failed to fetch document');
+  /// }
+  /// ```
+  /// **Next Steps**
+  /// - Upload a document to IPFS with [MotorFlutter.uploadDocument]
+  Future<GetDocumentResponse> fetchDocument({required String cid}) async {
+    final res = await MotorFlutterPlatform.instance.getDocument(GetDocumentRequest(
+      cid: cid,
+    ));
+    if (res == null) {
+      throw UnmarshalException<GetDocumentResponse>();
+    }
+    return res;
+  }
+
+  /// ### Refreshing a Bucket
+  ///
+  ///
+  refresh() async {
+    if (!_needsRefresh) {
+      return;
+    }
+    final ok = await MotorFlutterPlatform.instance.resolveBucket(did);
+    if (!ok) {
+      throw Exception('Failed to resolve bucket');
     }
     final contentList = await MotorFlutterPlatform.instance.getBucketObjects(did);
     final items = contentList.buckets.map((e) => BucketItem.fromBuffer(e.item)).toList();
@@ -84,23 +118,8 @@ class Bucket {
         docs.add(doc);
       }
     }
-    return docs;
-  }
-
-  /// Deletes a [SchemaDocument] from the Bucket given its [cid]
-  Future<bool> removeDocument(String cid) async {
-    if (!_isInitialized) {
-      await _resolve();
-    }
-    return await MotorFlutterPlatform.instance.removeBucketObject(did, cid);
-  }
-
-  // helper function to resolve the bucket
-  _resolve() async {
-    final ok = await MotorFlutterPlatform.instance.resolveBucket(did);
-    if (!ok) {
-      throw Exception('Failed to resolve bucket');
-    }
-    _isInitialized = true;
+    _cachedDocs.clear();
+    _cachedDocs.addAll(docs);
+    _lastUpdated = DateTime.now();
   }
 }