+++ lib/src/motor_flutter_base.dart
@@ -552,62 +552,6 @@ class MotorFlutter extends GetxService {
     return res;
   }
 
-  /// ### Retreiving a Document
-  ///
-  /// Fetches a document from IPFS using the provided [cid]. Returns the [SchemaDocument] if successful, and null if the document was not found.
-  ///
-  /// ```dart
-  /// final res = await MotorFlutter.to.getDocument('QmXyZ123');
-  /// if (res == null) {
-  ///   throw Exception('Failed to fetch document');
-  /// }
-  /// ```
-  /// **Next Steps**
-  /// - Upload a document to IPFS with [MotorFlutter.uploadDocument]
-  Future<GetDocumentResponse> getDocument({required String cid}) async {
-    final res = await MotorFlutterPlatform.instance.getDocument(GetDocumentRequest(
-      cid: cid,
-    ));
-    if (res == null) {
-      throw UnmarshalException<GetDocumentResponse>();
-    }
-    return res;
-  }
-
-  /// ### Upload a Document
-  ///
-  /// Uploads a document to IPFS. Returns the [UploadDocumentResponse] if successful, and null if the document was not found.
-  ///
-  /// ```dart
-  /// // Define a document
-  /// final def = SchemaDefinition(label: 'MySchema', fields: {'name': 'String', 'age': 'Int'});
-  ///
-  /// // Create empty document from definition
-  /// final doc = def.newDocument();
-  /// doc.set<String>('name', 'John');
-  /// doc.set<int>('age', 30);
-  ///
-  /// // Upload document to IPFS
-  /// final res = await MotorFlutter.to.uploadDocument(doc);
-  /// if (res == null) {
-  ///  throw Exception('Failed to upload document');
-  /// }
-  /// ```
-  /// **Next Steps**
-  /// - Get a document to IPFS with [MotorFlutter.getDocument]
-  Future<UploadDocumentResponse> uploadDocument({required SchemaDocument doc, required String label}) async {
-    final res = await MotorFlutterPlatform.instance.uploadDocument(UploadDocumentRequest(
-      schemaDid: doc.schemaDid,
-      document: doc.writeToBuffer(),
-      label: label,
-    ));
-    if (res == null) {
-      throw UnmarshalException<UploadDocumentResponse>();
-    }
-    res.document.cid = res.cid;
-    return res;
-  }
-
   /// ### Get Account Info
   ///
   /// Returns the current Accounts Info and updates the reactive Variables of [MotorFlutter].